requirements.txt
aenum==3.1.16
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
dataclasses-json==0.6.7
dotenv==0.9.9
geojson==3.2.0
geopandas==1.1.1
idna==3.10
imageio==2.37.0
lazy_loader==0.4
llvmlite==0.44.0
looseversion==1.3.0
marshmallow==3.26.1
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
osmnx==2.0.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pysheds==0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.4
requests-oauthlib==2.0.0
scikit-image==0.25.2
scipy==1.15.3
sentinelhub==3.11.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
tifffile==2025.6.1
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
utm==0.8.1
wheel==0.45.1
whitebox==2.3.6
geoserver-rest==2.9.1
